test2/kafka-docker/mongodb/requirements.txt
dnspython>=2.0.0
kafka-python==2.0.2
pymongo==4.4
zeep>=4.1.0